requirements.txt
build
duckdb
filelock
JPype1
protobuf
pyarrow
setuptools
twine
